Update and rename README.md to Project Book the software through artificial intelligence
The project is an artificial intelligence program that develops and writes software using artificial intelligence. The idea of the program is to write clear English words, and it turns it into a final program written in languages. It saves the application developers time and effort to write code, and a portion of it can be used to teach nos to code.